+++ src/components/user/login.spec.js
@@ -45,7 +45,7 @@ describe("Login works", () => {
     cy.get(".topbar")
       .find(".name")
       .should("have.text", "John Doe (ABC Tenant)");
-    cy.get(".dropdown-toggle").find(".dropdown-toggle").click();
+    cy.get(".sidebar-nav").find(".dropdown-toggle").click();
     cy.get("#DEF").click();
     cy.wait(["@groups-two"]);
     cy.get(".topbar")
@@ -108,6 +108,6 @@ describe("Login works", () => {
     cy.get("#login").click();
     cy.wait(["@token"]);
     cy.get(".name").should("have.text", "John Doe (ABC Tenant)");
-    cy.get(".sidebar-nav").find(".nav-item").should("have.length", 10);
+    cy.get(".sidebar-nav").find(".nav-item").should("have.length", 11);
   });
 });