+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#213](https://github.com/os2display/display-admin-client/pull/213)
+  add focus trap, to trap the keyboard focus in our modals
+  add user list
+  add user create
+  add user edit
+  add a bunch of todos
 - [#212](https://github.com/os2display/display-admin-client/pull/212)
   update color of login background
 - [#211](https://github.com/os2display/display-admin-client/pull/211)