+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 - [#214](https://github.com/os2display/display-admin-client/pull/214)
-  OIDC Code authorization flow
+  Use OIDC Code authorization flow and remove React StrictMode.
 - [#210](https://github.com/os2display/display-admin-client/pull/210)
   Use thumbnails for media list if they are set